sonar bug

+++ app/src/main/java/com/example/abilambin/nutritio/utils/Utils.java
@@ -16,7 +16,7 @@ public class Utils {
      * @return
      */
     public static int percent(float val, float obj){
-        if (obj == 0) return Math.round(val*100);
+        if (((int) obj ) == 0) return Math.round(val*100);
 
         float result = (val/obj)*100;
         return Math.round(result);
